main, service: move vesionbundle to pkg/project
